feat: experimental cryptography
