+++ packages/core/src/keystore/cryptography/ethers/keystore.ts
@@ -1,0 +1,110 @@
+/**
+ * Implements the JSON Keystore v3 Wallet encryption, decryption, and validation functionality.
+ */
+import { ethers } from 'ethers';
+import { type Keystore, type KeystoreAccount } from '../../types';
+import { assert, buildError, KEYSTORE } from '@vechain/sdk-errors';
+import { secp256k1 } from '../../../secp256k1';
+import { addressUtils } from '../../../address';
+import { Hex0x } from '../../../utils';
+import { SCRYPT_PARAMS } from './const';
+
+/**
+ * Encrypts a given private key into a keystore format using the specified password.
+ *
+ * @param privateKey - The private key to be encrypted.
+ * @param password - The password used for the encryption.
+ * @returns A Promise that resolves to the encrypted keystore.
+ */
+async function encrypt(
+    privateKey: Buffer,
+    password: string
+): Promise<Keystore> {
+    // Public and Address are derived from a private key
+    const derivePublicKey = secp256k1.derivePublicKey(privateKey);
+    const deriveAddress = addressUtils.fromPublicKey(
+        Buffer.from(derivePublicKey)
+    );
+
+    // Create keystore account compatible with ethers
+    const keystoreAccount: ethers.KeystoreAccount = {
+        address: deriveAddress,
+        privateKey: Hex0x.of(privateKey)
+    };
+
+    // Scrypt options
+    const encryptOptions: ethers.EncryptOptions = {
+        scrypt: {
+            N: SCRYPT_PARAMS.N,
+            r: SCRYPT_PARAMS.r,
+            p: SCRYPT_PARAMS.p
+        }
+    };
+
+    // Stirng version of keystore
+    const keystoreJsonString = await ethers.encryptKeystoreJson(
+        keystoreAccount,
+        password,
+        encryptOptions
+    );
+
+    return JSON.parse(keystoreJsonString) as Keystore;
+}
+
+/**
+ * Decrypts a keystore to obtain the private key using the given password.
+ *
+ * @throws{InvalidKeystoreError, InvalidKeystorePasswordError}
+ * @param keystore - The keystore containing the encrypted private key.
+ * @param password - The password used to decrypt the keystore.
+ * @returns A Promise that resolves to the decrypted KeystoreAccount or rejects if the keystore or password is invalid.
+ */
+async function decrypt(
+    keystore: Keystore,
+    password: string
+): Promise<KeystoreAccount> {
+    // Invalid keystore
+    assert(
+        'keystore.decrypt',
+        isValid(keystore),
+        KEYSTORE.INVALID_KEYSTORE,
+        'Invalid keystore. Ensure the keystore is properly formatted and contains the necessary data.',
+        {
+            keystore
+        }
+    );
+
+    try {
+        return (await ethers.decryptKeystoreJson(
+            JSON.stringify(keystore),
+            password
+        )) as KeystoreAccount;
+    } catch (e) {
+        throw buildError(
+            'keystore.decrypt',
+            KEYSTORE.INVALID_PASSWORD,
+            'Decryption failed: Invalid Password for the given keystore.',
+            {
+                keystore,
+                password
+            },
+            e
+        );
+    }
+}
+
+/**
+ * Validates if the provided keystore adheres to the expected format and structure.
+ *
+ * @param keystore - The keystore to be validated.
+ * @returns A boolean indicating whether the keystore is valid or not.
+ */
+function isValid(keystore: Keystore): boolean {
+    return ethers.isKeystoreJson(JSON.stringify(keystore));
+}
+
+/**
+ * Exports the keystore functions for encryption, decryption, and validation.
+ */
+const keystore = { encrypt, decrypt, isValid };
+export { keystore };
