+++ packages/core/src/keystore/types.d.ts
@@ -1,28 +1,5 @@
 /**
- * Represents a KeyStore object, which contains encrypted information about a cryptocurrency wallet
- * according [Web3 Secret Storage Definition](https://ethereum.org/en/developers/docs/data-structures-and-encoding/web3-secret-storage/).
- *
- * @interface KeyStore
- * @property {string} address - The address of the cryptocurrency wallet.
- * @property {Object} crypto - The object containing the encrypted data.
- * @property {Cipher} crypto.cipher - The cipher algorithm used for encryption.
- * @property {Object} crypto.cipherparams - The parameters for the cipher algorithm.
- * @property {string} crypto.cipherparams.iv - The IV (Initialization Vector) used for encryption.
- * @property {string} crypto.ciphertext - The encrypted data.
- * @property {string} crypto.kdf - The key derivation function used for encryption.
- * @property {Object} crypto.kdfparams - The parameters for the key derivation function.
- * @property {number} crypto.kdfparams.dklen - The length of the derived key.
- * @property {number} crypto.kdfparams.n - The CPU/memory cost parameter for the key derivation function.
- * @property {number} crypto.kdfparams.p - The parallelization parameter for the key derivation function.
- * @property {number} crypto.kdfparams.r - The block size parameter for the key derivation function.
- * @property {string} crypto.kdfparams.salt - The salt used for the key derivation function.
- * @property {string} crypto.mac - The message authentication code of the encrypted data.
- * @property {string} id - The unique identifier of the KeyStore object.
- * @property {number} version - The version number of the KeyStore object.
- */
-
-/**
- * @interface KeyStore
+ * @interface Keystore
  * Represents a
  * [Web3 Secret Storage](https://ethereum.org/en/developers/docs/data-structures-and-encoding/web3-secret-storage)
  * keystore object that holds information about a private cryptographic key.
@@ -48,7 +25,7 @@
  * for the key store.
  * @property {number} version - The version number of the key store.
  */
-interface KeyStore {
+interface Keystore {
     address: string;
     crypto: {
         cipher: string;
@@ -90,7 +67,13 @@ interface KeyStore {
  */
 interface KeystoreAccount {
     address: string;
-    privateKey: Uint8Array;
+    privateKey: string;
+    // @NOTE: Added ONLY for compatibility with ethers KeystoreAccount of ethers.
+    mnemonic?: {
+        path?: string;
+        locale?: string;
+        entropy: string;
+    };
 }
 
-export { type KeyStore, type KeystoreAccount };
+export { type Keystore, type KeystoreAccount };
